Bug: Change th2m to t2m in anavinfo
The latest GSI code used in RRFS changed the name of 2-m temperature in the GSI bundle from th2m to t2m to reflect the fact that this is sensible instead of potential temperature. In order to prevent a crash in GSI, anavinfo must be updated accordingly. Now WRF_FCST_OSSE can be used with the latest GSI code.