Fix GSI diag namelists
Updated namelists differ from the HRRR because the underlying executable is different. The GSI diag reader used here comes from GSI-Utils and creates text output (unlike RRFS, which creates netCDF4 output). The text output is very similar to the HRRR except that it is missing the final column (I do not know what this column is for, as it is not included in appendix A of DTC's GSI documentation). GSI satellits diagnostics are also turned off because there is no satellite DA and the obs counter is also turned off because I could not find the executable.