Add back ungrib.exe on Hercules
I don't think this executable actually works on Hercules, but if some
executable is not specified, then the ungrib tasks will crash. This
executable should never be run because all necessary fields have
already been ungribbed and are saved on MSU HPC2.

From what I recall, we do not have a working version of ungrib.exe on
Hercules.